requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.0
gunicorn==21.2.0
python-decouple==3.8
django-cors-headers==4.3.1
whitenoise==6.6.0

